Fix ESLint in botframework-streaming
