+++ libraries/botframework-streaming/eslint.config.cjs
@@ -1,13 +1,8 @@
-const onlyWarn = require("eslint-plugin-only-warn");
-const sharedConfig = require("../../eslint.config.cjs")
+const sharedConfig = require('../../eslint.config.cjs');
 
 module.exports = [
     ...sharedConfig,
     {
-        ignores: ["**/es5/"],
+        ignores: ['**/es5/'],
     },
-    {
-        plugins: {
-            "only-warn": onlyWarn,
-        },
-    }]
+];
