+++ libraries/botframework-streaming/src/interfaces/INodeDuplex.ts
@@ -1,4 +1,3 @@
-/* eslint-disable @typescript-eslint/no-explicit-any, @typescript-eslint/ban-types */
 /**
  * @module botframework-streaming
  */
