+++ libraries/botframework-streaming/src/interfaces/INodeSocket.ts
@@ -1,4 +1,3 @@
-/* eslint-disable @typescript-eslint/ban-types */
 /**
  * @module botframework-streaming
  */
