+++ libraries/botframework-streaming/src/subscribableStream.ts
@@ -14,7 +14,6 @@ export class SubscribableStream extends Duplex {
     length = 0;
 
     private readonly bufferList: Buffer[] = [];
-
     private _onData: (chunk: any) => void;
 
     /**
@@ -34,7 +33,6 @@ export class SubscribableStream extends Duplex {
      * @param _encoding The encoding. Unused in the implementation of Duplex.
      * @param callback Callback for when this chunk of data is flushed.
      */
-
     _write(chunk: any, _encoding: string, callback: (error?: Error | null) => void): void {
         const buffer = Buffer.from(chunk);
         this.bufferList.push(buffer);
@@ -72,7 +70,6 @@ export class SubscribableStream extends Duplex {
      *
      * @param onData Callback to be called when onData is executed.
      */
-
     subscribe(onData: (chunk: any) => void): void {
         this._onData = onData;
     }
