+++ libraries/botframework-streaming/tests/ContentStream.test.js
@@ -67,7 +67,7 @@ describe('Streaming Extensions ContentStream Tests ', function () {
     it('does not return the stream when it is is undefined', function () {
         const cs = new ContentStream('1', new PayloadAssembler(streamManager, { id: 'csa1', header: header }));
 
-        expect(cs.getStream()).to.not.be.undefined;
+        expect(cs.getStream()).to.not.equal(undefined);
     });
 
     it('reads a stream of length 0 and returns an empty string', function () {
@@ -99,7 +99,7 @@ describe('Streaming Extensions ContentStream Tests ', function () {
         return cs
             .readAsJson()
             .then((data) => {
-                expect(data).to.not.be.undefined;
+                expect(data).to.not.be.undefined();
             })
             .catch((err) => {
                 expect(err.toString()).to.equal('SyntaxError: Unexpected end of JSON input');
